+++ app/src/main/java/com/fibelatti/photowidget/model/PhotoWidgetTapAction.kt
@@ -9,7 +9,8 @@ enum class PhotoWidgetTapAction(
 
     NONE(label = R.string.photo_widget_configure_tap_action_none),
     VIEW_FULL_SCREEN(label = R.string.photo_widget_configure_tap_action_view_full_screen),
-    VIEW_NEXT_PHOTO(label = R.string.photo_widget_configure_tap_action_next_photo),
+    VIEW_IN_GALLERY(label = R.string.photo_widget_configure_tap_action_view_in_gallery),
+    VIEW_NEXT_PHOTO(label = R.string.photo_widget_configure_tap_action_flip_photo),
     APP_SHORTCUT(label = R.string.photo_widget_configure_tap_action_app_shortcut),
     ;
 }