+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Make it possible to view the previous photo on tap
+* Add "View in gallery" as a tap action
 
 ## [v1.9.1] - 2024-06-11